+++ voting/frontend/app/voters/page.tsx
@@ -3,9 +3,9 @@
 import IsConnected from "@/components/IsConnected";
 import GetVoter from "@/components/voters/GetVoter";
 import ProposalManager from "@/components/voters/ProposalManager";
-import SendVote from "@/components/voters/SendVote";
 import Event from "@/components/Event";
 import {WorkflowStepperManager} from "@/components/home/WorkflowStepperManager";
+import * as React from "react";
 
 
 export default function Voters() {
@@ -15,7 +15,6 @@ export default function Voters() {
         <WorkflowStepperManager />
         <GetVoter />
         <ProposalManager />
-        {/*<SendVote/>*/}
 
         <div className="mt-2">
             <Event name='ProposalRegistered'></Event>