+++ voting/frontend/components/voters/SendProposal.tsx
@@ -1,14 +1,19 @@
 "use client"
 
 import { writeContractByFunctionName } from "@/utils"
-import {useEffect, useState} from "react"
+import { useState } from "react"
 import { useToast } from "@chakra-ui/react"
-import SendVote from "@/components/voters/SendVote";
+import { useContractEvent } from "wagmi";
+import { abi, contractAddress } from "@/constants";
+import { Log } from "viem";
+import GetProposalsList from "@/components/voters/GetProposalsList";
 
 const SendProposal = () => {
+
     const [proposal, setProposal] = useState('')
     const toast = useToast()
-    const [proposalCreated, setProposalCreated] = useState(false)
+    const [logs, setLogs] = useState<Log[]>()
+    const [ newProposal, setNewProposal] = useState(false)
 
     const submitProposal = () => {
         if (proposal.trim().length <= 0) {
@@ -22,6 +27,7 @@ const SendProposal = () => {
         } else {
             writeContractByFunctionName('addProposal', proposal).then(
                 () => {
+                    setNewProposal(true)
                     toast({
                         title: 'Proposal successfully added.',
                         description: `proposal: ${proposal}.`,
@@ -29,7 +35,6 @@ const SendProposal = () => {
                         duration: 5000,
                         isClosable: true,
                     })
-                    setProposalCreated(true)
                 }
             ).catch(
                 err => toast({
@@ -39,13 +44,26 @@ const SendProposal = () => {
                     duration: 5000,
                     isClosable: true,
                 })
-            )
+            ).finally(() => {
+                setProposal("")
+
+            })
         }
-        setProposal("")
-        setProposalCreated(false)
     }
 
+    useContractEvent({
+        address: contractAddress,
+        abi: abi,
+        eventName: 'ProposalRegistered',
+        listener(log) {
+            setLogs(log)
+            setNewProposal(false)
+        }
+    })
+
+
     return (<>
+
         <section className="m-2 mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-800 text-indigo-100 shadow-lg drop-shadow-lg border-indigo-600 border">
             <h2 className="font-bold text-lg text-center mb-3">Send a Proposal</h2>
             <form className="w-full m-auto p-3 pr-3">
@@ -66,9 +84,9 @@ const SendProposal = () => {
                     </button>
                 </div>
             </form>
-        </section>
+            <GetProposalsList addNewProposal={newProposal}/>
 
-        <SendVote newProposal={proposalCreated} />
+        </section>
     </>);
 }
 export default SendProposal;